# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KABBAK API",
    "version": "0.2.0",
    "description": "Read-only API backend for the KABBAK frontend. All content responses use a {data, meta?} envelope (see README). Protected routes accept x-api-key or Authorization: Bearer <key>. Named clients can carry a basic, premium, or pro+ accessLevel, plus optional roles and scopes for privileged admin routes.",
    "x-service-name": "kabbak-api"
  },
  "servers": [
    {
      "url": "http://localhost:3100/api/v1",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/calendar/months": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "List calendar months reference entries",
        "operationId": "listCalendarMonths",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/calendar/months/{monthId}": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get one calendar month reference entry",
        "operationId": "getCalendarMonth",
        "parameters": [
          {
            "$ref": "#/components/parameters/MonthId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/calendar/holidays": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "List holiday reference entries",
        "operationId": "listCalendarHolidays",
        "parameters": [
          {
            "$ref": "#/components/parameters/HolidayKind"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/calendar/holidays/{holidayId}": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get one holiday reference entry",
        "operationId": "getCalendarHoliday",
        "parameters": [
          {
            "$ref": "#/components/parameters/HolidayId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/astrology/natal": {
      "get": {
        "tags": [
          "astrology"
        ],
        "summary": "Cast a natal chart for a birth date, optional time, and location",
        "operationId": "getNatalChart",
        "parameters": [
          {
            "$ref": "#/components/parameters/Latitude"
          },
          {
            "$ref": "#/components/parameters/Longitude"
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Birth date YYYY-MM-DD. Required unless datetime is set."
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Birth time HH:MM or HH:MM:SS. Interpreted as local mean time at longitude unless utcOffsetMinutes is set."
          },
          {
            "name": "datetime",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "ISO instant. Overrides date/time when present."
          },
          {
            "name": "timeUnknown",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "utcOffsetMinutes",
            "in": "query",
            "required": false,
            "schema": { "type": "number" }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "placeId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/astrology/planets": {
      "get": {
        "tags": [
          "astrology"
        ],
        "summary": "List astrology planets",
        "operationId": "listPlanets",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/astrology/planets/{planetId}": {
      "get": {
        "tags": [
          "astrology"
        ],
        "summary": "Get one planet entry",
        "operationId": "getPlanet",
        "parameters": [
          {
            "$ref": "#/components/parameters/PlanetId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/astrology/signs": {
      "get": {
        "tags": [
          "astrology"
        ],
        "summary": "List zodiac signs",
        "operationId": "listSigns",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/astrology/signs/{signId}": {
      "get": {
        "tags": [
          "astrology"
        ],
        "summary": "Get one zodiac sign",
        "operationId": "getSign",
        "parameters": [
          {
            "$ref": "#/components/parameters/SignId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/astrology/decans": {
      "get": {
        "tags": [
          "astrology"
        ],
        "summary": "List astrology decans",
        "operationId": "listDecans",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/astrology/decans/{decanId}": {
      "get": {
        "tags": [
          "astrology"
        ],
        "summary": "Get one astrology decan",
        "operationId": "getDecan",
        "parameters": [
          {
            "$ref": "#/components/parameters/DecanId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/iching": {
      "get": {
        "tags": [
          "iching"
        ],
        "summary": "Get the I Ching dataset",
        "operationId": "getIChing",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/iching/hexagrams/{number}": {
      "get": {
        "tags": [
          "iching"
        ],
        "summary": "Get one hexagram",
        "operationId": "getHexagram",
        "parameters": [
          {
            "$ref": "#/components/parameters/HexagramNumber"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/iching/trigrams/{name}": {
      "get": {
        "tags": [
          "iching"
        ],
        "summary": "Get one trigram",
        "operationId": "getTrigram",
        "parameters": [
          {
            "$ref": "#/components/parameters/TrigramName"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/correspondences": {
      "get": {
        "tags": ["correspondences"],
        "summary": "List correspondence entity kinds and counts",
        "operationId": "listCorrespondenceKinds",
        "responses": {
          "200": { "$ref": "#/components/responses/JsonPayload" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/correspondences/{kind}/{id}": {
      "get": {
        "tags": ["correspondences"],
        "summary": "Get one correspondence entity and its relations",
        "operationId": "getCorrespondence",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonPayload" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/alphabets": {
      "get": {
        "tags": [
          "alphabets"
        ],
        "summary": "List alphabet datasets",
        "operationId": "listAlphabets",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/alphabets/{scriptId}": {
      "get": {
        "tags": [
          "alphabets"
        ],
        "summary": "Get one alphabet dataset",
        "operationId": "getAlphabet",
        "parameters": [
          {
            "$ref": "#/components/parameters/ScriptId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/gods": {
      "get": {
        "tags": [
          "gods"
        ],
        "summary": "List god datasets",
        "operationId": "listGodGroups",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/gods/{groupId}": {
      "get": {
        "tags": [
          "gods"
        ],
        "summary": "Get one god group or alias match",
        "operationId": "getGodGroup",
        "parameters": [
          {
            "$ref": "#/components/parameters/GodGroupId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/chakras": {
      "get": {
        "tags": [
          "chakras"
        ],
        "summary": "List chakra datasets",
        "operationId": "listChakras",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/chakras/{chakraId}": {
      "get": {
        "tags": [
          "chakras"
        ],
        "summary": "Get one chakra entry",
        "operationId": "getChakra",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChakraId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/enochian": {
      "get": {
        "tags": [
          "enochian"
        ],
        "summary": "List Enochian datasets",
        "operationId": "listEnochianGroups",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/enochian/{groupId}": {
      "get": {
        "tags": [
          "enochian"
        ],
        "summary": "Get one Enochian group",
        "operationId": "getEnochianGroup",
        "parameters": [
          {
            "$ref": "#/components/parameters/EnochianGroupId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ResponseMeta": {
        "type": "object",
        "description": "Optional metadata attached to success responses for pagination, tracing and versioning.",
        "properties": {
          "requestId": {
            "type": "string",
            "description": "Unique identifier for the request, echoed from x-request-id or generated."
          },
          "version": {
            "type": "string",
            "description": "API/service version at the time of the response. Allows GUIs to detect shape changes."
          },
          "total": {
            "type": "integer",
            "description": "Total number of items available (for list/paginated responses)."
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "count": {
            "type": "integer",
            "description": "Number of items in this response page."
          }
        }
      },
      "ApiSuccess": {
        "type": "object",
        "description": "Standard success envelope for all content API responses. GUIs should expect data + optional meta.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "description": "Primary payload. May be object, array, or primitive depending on endpoint."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "additionalProperties": false
      },
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "message",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "description": "Optional structured error details.",
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
...
    },
    "responses": {
      "JsonPayload": {
        "description": "Standard JSON success response using the ApiSuccess envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiSuccess"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "A valid API key was not supplied.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "The request parameters could not be processed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "Latitude": {
        "name": "latitude",
        "in": "query",
        "required": false,
        "schema": {
          "type": "number"
        }
      },
      "Longitude": {
        "name": "longitude",
        "in": "query",
        "required": false,
        "schema": {
          "type": "number"
        }
      },
...
      "PlanetId": {
        "name": "planetId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "SignId": {
        "name": "signId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "DecanId": {
        "name": "decanId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "MonthId": {
        "name": "monthId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "HolidayId": {
        "name": "holidayId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "HolidayKind": {
        "name": "kind",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "all",
            "celestial",
            "calendar"
          ]
        }
      },
      "HexagramNumber": {
        "name": "number",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "TrigramName": {
        "name": "name",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "ScriptId": {
        "name": "scriptId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "GodGroupId": {
        "name": "groupId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "ChakraId": {
        "name": "chakraId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "EnochianGroupId": {
        "name": "groupId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    }
  }
}